LayoutTests/platform/chromium-mac/editing/input/ime-candidate-window-position.html

<html>
<head>
<style>
* {
    padding: 0px;
    margin: 0px;
    border: 0px;
}
</style>
<link rel="stylesheet" href="../../../../fast/js/resources/js-test-style.css">
<script src="../../../../fast/js/resources/js-test-pre.js"></script>
<script>
window.jsTestIsAsync = true;
var successfullyParsed = false;

function doTest() {
    if (window.layoutTestController) {
        frame = document.getElementsByTagName('iframe')[0];
        input = frames['iframe'].document.getElementsByTagName('input')[0];
        input.focus();
        rect = textInputController.firstRectForCharacterRange(0, 0);
        debug('rect[0] denotes x coodinate and rect[1] denotes y coodinate');
        // I'm not sure why we need to +1 for x coodinate here but
        // firstRectForCharacterRange returns that value.
        shouldBe('rect[0]', 'frame.offsetLeft + input.offsetLeft + 1');
        shouldBe('rect[1]', 'frame.offsetTop + input.offsetTop');
    }
    successfullyParsed = true;
    finishJSTest();
}
</script>
</head>
<body onload="doTest()">
<div>
This test ensures that the IME candidate window appears at the appropriate position.  Enter some text with IME in the following gray box and convert them.  Then check whether the candidate window appears the below of the composition text.  In DRT, this test checks it by calling textInputController.firstRectForCharacterRange().
</div>
<div id="console"></div>
<iframe name="iframe" style="position: absolute; top: 200px; left: 200px;" src="../resources/ime-candidate-window-position-iframe.html" width="300" height="300">
</iframe>
<script src="../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
